LKE-9831 Mention users in case comments (#644)
* LKE-9831 Mention users in case comments

* Update types.ts

* Update types.ts

* Update types.ts

* LKE-9831 post review

* run lint

* LKE-9846: Notify users mentioned in cases (#646)

* LKE-9846 update email notification config

* Update types.ts

* Update types.ts
