+++ src/api/Alert/types.ts
@@ -5,6 +5,7 @@
  */
 
 import {
+  CommentMention,
   ICurrencyOptions,
   IDataSourceParams,
   IGetSubGraphParams,
@@ -299,7 +300,12 @@ export interface CaseAction extends PersistedItem {
   user: Pick<User, 'id' | 'username' | 'email'>;
   action: CaseActionType;
   comment?: string;
+  metadata?: CaseActionMetadata;
   assignedUser?: Pick<User, 'id' | 'username' | 'email'>;
+}
+
+export interface CaseActionMetadata {
+  mentions: CommentMention[];
 }
 
 export interface IDoCaseActionParams extends IDataSourceParams {
