+++ src/api/Config/types.ts
@@ -427,11 +427,17 @@ export interface ISMTPAuthConfig {
 }
 
 export interface IEmailNotificationsConfig {
+  // Alert notifications.
   alertNotifications: boolean;
-  visualizationNotifications: boolean;
   newCasesDigestNotificationFrequency: string;
   caseAssignmentNotificationFrequency: string;
+  caseMentionNotificationFrequency: string;
+
+  // Visualization notifications.
+  visualizationNotifications: boolean;
   visualizationMentionNotificationFrequency: string;
+
+  // Email configuration.
   mailer: IMailerConfig;
   fromEmail: string;
 }
