+++ src/api/User/types.ts
@@ -157,12 +157,3 @@ export interface IGetAssetTransferEligibleUsersParams {
 export interface ICountSharedAssets {
   count: number;
 }
-
-export interface DeletableUser {
-  username: string | DeletedUser;
-  email: string | DeletedUser;
-}
-
-export interface MentionedUser extends DeletableUser {
-  hasAccess: boolean;
-}
