+++ src/api/commonTypes.ts
@@ -3,6 +3,7 @@
  *
  * - Created on 2019-09-30.
  */
+import {DeletedUser} from '../../src/api/User/types';
 
 export interface GenericObject<T = unknown> {
   [key: string]: T;
@@ -91,4 +92,19 @@ export interface ICurrencyOptions {
   type: 'currency';
   format: CurrencyFormat;
   symbol?: string;
+}
+
+export interface CommentMention {
+  position: number;
+  length: number;
+  user: MentionedUser;
+}
+
+export interface DeletableUser {
+  username: string | DeletedUser;
+  email: string | DeletedUser;
+}
+
+export interface MentionedUser extends DeletableUser {
+  hasAccess: boolean;
 }
